requirements-backend.txt
# Cloud Backend Dependencies
# Install when building Flask/FastAPI backend

flask>=3.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
pyjwt>=2.8.0
passlib>=1.7.4
email-validator>=2.1.0
redis>=5.0.0
requests>=2.31.0
python-socketio>=5.11.0
python-engineio>=4.8.0
httpx>=0.27.0
